backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
anthropic
resend
pdfplumber
python-docx
python-dotenv
